created useGuestUser hooks
